# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReminderCreate": {
        "properties": {
          "due_in_seconds": {
            "default": 0,
            "maximum": 2592000.0,
            "minimum": 0.0,
            "title": "Due In Seconds",
            "type": "integer"
          },
          "form_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Form Id"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "retry_delay_seconds": {
            "default": 3600,
            "maximum": 2592000.0,
            "minimum": 60.0,
            "title": "Retry Delay Seconds",
            "type": "integer"
          },
          "title": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "ReminderCreate",
        "type": "object"
      },
      "ReminderResponse": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "delivery_count": {
            "title": "Delivery Count",
            "type": "integer"
          },
          "delivery_retry_delay_seconds": {
            "title": "Delivery Retry Delay Seconds",
            "type": "integer"
          },
          "enqueue_status": {
            "default": "pending",
            "enum": [
              "pending",
              "queued",
              "failed"
            ],
            "title": "Enqueue Status",
            "type": "string"
          },
          "enqueued_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enqueued At"
          },
          "form_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Form Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_delivered_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Delivered At"
          },
          "last_enqueue_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Enqueue Error"
          },
          "next_run_at": {
            "format": "date-time",
            "title": "Next Run At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "retry_delay_seconds": {
            "title": "Retry Delay Seconds",
            "type": "integer"
          },
          "skip_count": {
            "title": "Skip Count",
            "type": "integer"
          },
          "status": {
            "enum": [
              "pending",
              "skipped",
              "completed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "form_id",
          "title",
          "payload",
          "status",
          "retry_delay_seconds",
          "delivery_retry_delay_seconds",
          "next_run_at",
          "skip_count",
          "delivery_count",
          "created_at",
          "updated_at"
        ],
        "title": "ReminderResponse",
        "type": "object"
      },
...
      "ReminderSkipRequest": {
        "properties": {
          "retry_delay_seconds": {
            "anyOf": [
              {
                "maximum": 2592000.0,
                "minimum": 60.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Delay Seconds"
          }
        },
        "title": "ReminderSkipRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "FastAPI with PostgreSQL database.",
    "title": "ping me Project API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/reminders": {
      "get": {
        "operationId": "list_reminders_reminders_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "pending",
                      "skipped",
                      "completed",
                      "cancelled"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "form_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Form Id"
            }
          },
          {
            "in": "query",
            "name": "due_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Due Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReminderResponse"
                  },
                  "title": "Response List Reminders Reminders Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Reminders",
        "tags": [
          "Reminders"
        ]
      },
      "post": {
        "operationId": "create_reminder_reminders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Reminder",
        "tags": [
          "Reminders"
        ]
      }
    },
    "/reminders/current": {
      "get": {
        "operationId": "get_current_reminders_reminders_current_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReminderResponse"
                  },
                  "title": "Response Get Current Reminders Reminders Current Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Current Reminders",
        "tags": [
          "Reminders"
        ]
      }
    },
    "/reminders/{reminder_id}/cancel": {
      "post": {
        "operationId": "cancel_reminder_reminders__reminder_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "reminder_id",
            "required": true,
            "schema": {
              "title": "Reminder Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Cancel Reminder",
        "tags": [
          "Reminders"
        ]
      }
    },
    "/reminders/{reminder_id}/complete": {
      "post": {
        "operationId": "complete_reminder_reminders__reminder_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "reminder_id",
            "required": true,
            "schema": {
              "title": "Reminder Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Complete Reminder",
        "tags": [
          "Reminders"
        ]
      }
    },
    "/reminders/{reminder_id}/skip": {
      "post": {
        "operationId": "skip_reminder_reminders__reminder_id__skip_post",
        "parameters": [
          {
            "in": "path",
            "name": "reminder_id",
            "required": true,
            "schema": {
              "title": "Reminder Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReminderSkipRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Skip Reminder",
        "tags": [
          "Reminders"
        ]
      }
    },
...
  }
}